+++ ignite/services/plugin/scaffold.go
@@ -6,11 +6,15 @@ import (
 	"os"
 	"path"
 	"path/filepath"
+	"strings"
 
 	"github.com/gobuffalo/genny/v2"
 	"github.com/gobuffalo/plush/v4"
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
 
 	"github.com/ignite/cli/v28/ignite/pkg/errors"
+	"github.com/ignite/cli/v28/ignite/pkg/gocmd"
 	"github.com/ignite/cli/v28/ignite/pkg/xgenny"
 )
 
@@ -21,6 +25,7 @@ var fsPluginSource embed.FS
 func Scaffold(ctx context.Context, dir, moduleName string, sharedHost bool) (string, error) {
 	var (
 		name     = filepath.Base(moduleName)
+		title    = toTitle(name)
 		finalDir = path.Join(dir, name)
 		g        = genny.New()
 		template = xgenny.NewEmbedWalker(
@@ -42,6 +47,7 @@ func Scaffold(ctx context.Context, dir, moduleName string, sharedHost bool) (str
 	pctx := plush.NewContextWithContext(ctx)
 	pctx.Set("ModuleName", moduleName)
 	pctx.Set("Name", name)
+	pctx.Set("Title", title)
 	pctx.Set("SharedHost", sharedHost)
 
 	g.Transformer(xgenny.Transformer(pctx))
@@ -55,5 +61,16 @@ func Scaffold(ctx context.Context, dir, moduleName string, sharedHost bool) (str
 		return "", errors.WithStack(err)
 	}
 
+	if err := gocmd.ModTidy(ctx, finalDir); err != nil {
+		return "", errors.WithStack(err)
+	}
+	if err := gocmd.Fmt(ctx, finalDir); err != nil {
+		return "", errors.WithStack(err)
+	}
+
 	return finalDir, nil
+}
+
+func toTitle(s string) string {
+	return strings.ReplaceAll(strings.ReplaceAll(cases.Title(language.English).String(s), "_", ""), "-", "")
 }